emtech2_project/requirements.txt
tensorflow
opencv-python
pillow
numpy
streamlit
scikit-learn
matplotlib
